feat: 成果物・合格バッジもレビューと同じく専用ページへ遷移に統一
トップ統計タイルの挙動を3つとも「専用ページへ遷移」に揃えた。
成果物→/works、合格バッジ→/works?approved=1、レビュー→/reviews。

- 一覧UI（絞り込み/並び替え＋カードグリッド）を WorksList 部品に抽出し、
  ランディングの WORKS セクションと専用ページ WorksPage で共有（重複排除）
- WorksPage は ?approved=1 で合格作品のみに絞った状態で開き、見出し・
  説明文も切り替える。検索ボックスも備える
- トップの WORKS 一覧・検索・観点カテゴリ導線は従来どおり維持

backend 変更なし（既存 GET /api/posts?approved=true を再利用）。
frontend build／lint green。
